[AssetEnqueur] create a DifferentialServingAssetEnqueur
